+++ Changelog.md
@@ -14,11 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add a feature to allow deskelectronjs engine to change page base on internel GUI url. Completed on 2025-04-24
+- Add download object to handler webview options. Completed on 2025-05-15
+- Add download function to webnodejs engine. Completed on 2025-05-15
 
 ### Changed
 
 - Upgrade framework package.json dependencies. Completed on 2025-04-24
-- Change success.log from webnodejs engine which will only accept by project api and gui request. Completed on 2025-05-04
 
 ### Deprecated